Restructured the format
