+++ projects/customer_support_chatbot_using_langgraph/src/utils/tools.py
@@ -3,8 +3,72 @@ import numpy as np
 import openai
 from langchain_core.tools import tool
 from dotenv import load_dotenv
+import sqlite3
+from datetime import date, datetime
+from typing import Optional
+import os, shutil, sqlite3
+import pandas as pd
+import requests
+import pytz
+from langchain_core.runnables import RunnableConfig
 
 load_dotenv()
+
+
+db_url = "https://storage.googleapis.com/benchmarks-artifacts/travel-db/travel2.sqlite"
+local_file = "conf/travel2.sqlite"
+backup_file = "conf/travel2.backup.sqlite"
+overwrite = False
+
+if overwrite or not os.path.exists(local_file):
+    response = requests.get(db_url)
+    response.raise_for_status()
+    with open(local_file, "wb") as f:
+        f.write(response.content)
+    shutil.copy(local_file, backup_file)
+
+
+def update_dates(file):
+    shutil.copy(backup_file, file)
+    conn = sqlite3.connect(file)
+    cursor = conn.cursor()
+    tables = pd.read_sql(
+        "SELECT name FROM sqlite_master WHERE type='table';", conn
+    ).name.tolist()
+    tdf = {}
+    for t in tables:
+        tdf[t] = pd.read_sql(f"SELECT * from {t}", conn)
+
+    example_time = pd.to_datetime(
+        tdf["flights"]["actual_departure"].replace("\\N", pd.NaT)
+    ).max()
+    current_time = pd.to_datetime("now").tz_localize(example_time.tz)
+    time_diff = current_time - example_time
+    tdf["bookings"]["book_date"] = (
+        pd.to_datetime(tdf["bookings"]["book_date"].replace("\\N", pd.NaT), utc=True)
+        + time_diff
+    )
+
+    datetime_columns = [
+        "scheduled_departure",
+        "scheduled_arrival",
+        "actual_departure",
+        "actual_arrival",
+    ]
+    for column in datetime_columns:
+        tdf["flights"][column] = (
+            pd.to_datetime(tdf["flights"][column].replace("\\N", pd.NaT)) + time_diff
+        )
+
+    for table_name, df in tdf.items():
+        df.to_sql(table_name, conn, if_exists="replace", index=False)
+    del df
+    del tdf
+    conn.commit()
+    conn.close()
+    return file
+
+db = update_dates(local_file)
 
 response = requests.get(
     "https://storage.googleapis.com/benchmarks-artifacts/travel-db/swiss_faq.md"
@@ -51,3 +115,4 @@ def lookup_policy(query: str) -> str:
     Use this before making any flight changes performing other 'write' events."""
     docs = retriever.query(query, k=2)
     return "\n\n".join([doc["page_content"] for doc in docs])
+
